Rename jrdev/ui/textual to jrdev/ui/tui
