+++ src/jrdev/ui/tui/bordered_switcher.py
