+++ src/jrdev/ui/tui/chat_input_widget.py
@@ -2,7 +2,7 @@ from textual.color import Color
 from typing import Optional
 import logging
 
-from jrdev.ui.textual.input_widget import CommandTextArea
+from jrdev.ui.tui.input_widget import CommandTextArea
 
 logger = logging.getLogger("jrdev")
