+++ src/jrdev/ui/tui/chat_view_widget.py
@@ -9,11 +9,11 @@ from textual.color import Color
 from typing import Optional
 import logging
 
-from jrdev.ui.textual.command_request import CommandRequest
+from jrdev.ui.tui.command_request import CommandRequest
 from jrdev.ui.textual_events import TextualEvents
-from jrdev.ui.textual.chat_input_widget import ChatInputWidget
+from jrdev.ui.tui.chat_input_widget import ChatInputWidget
 from jrdev.messages.thread import MessageThread, USER_INPUT_PREFIX
-from jrdev.ui.textual.message_bubble import MessageBubble
+from jrdev.ui.tui.message_bubble import MessageBubble
 
 logger = logging.getLogger("jrdev")
