+++ src/jrdev/ui/tui/file_deletion_screen.py
