+++ src/jrdev/ui/tui/model_selection_widget.py
