+++ src/jrdev/ui/tui/terminal_text_area.py
